minicpm_omni_requirements.txt
Pillow>=10.1.0
torch>=2.3.1
torchaudio>=2.3.1
torchvision>=0.18.1
transformers>=4.44.2
librosa
soundfile>=0.12.1
vector-quantize-pytorch>=1.18.5
vocos>=0.1.0
decord
moviepy
